refactor(health): rename health module to checkHealth and centralize route imports
