+++ src/infra/http/plugins/__test__/rate-limit.test.ts
@@ -1,7 +1,7 @@
 import APP from '@infra/http/app';
 
 import { rateLimitPlugin } from '@plugins/rate-limit';
-import healthRoutes from '@modules/health/router/health.router';
+import healthRoutes from '@modules/health/health.routes';
 
 describe('Rate limiting', () => {
   let app: APP;
