+++ src/modules/health/useCase/checkHealth/checkHealth.controller.ts
@@ -1,0 +1,5 @@
+import { CheckHealthResponse } from '@modules/health/useCase/checkHealth/checkHealth.schema';
+
+export async function checkHealthController(): Promise<CheckHealthResponse> {
+  return { status: 'ok', uptime: process.uptime() };
+}
