+++ src/modules/health/useCase/checkHealth/checkHealth.schema.ts
@@ -1,0 +1,16 @@
+export interface CheckHealthResponse {
+  status: string;
+  uptime: number;
+}
+
+export const checkHealthSchema = {
+  response: {
+    200: {
+      type: 'object',
+      properties: {
+        status: { type: 'string' },
+        uptime: { type: 'number' },
+      },
+    },
+  },
+};
