+++ src/modules/modules.routers.ts
@@ -1,0 +1,7 @@
+import { FastifyInstance } from 'fastify';
+
+import healthRoutes from '@src/modules/health/health.routes';
+
+export default async function moduleRoutes(app: FastifyInstance) {
+  app.register(healthRoutes, { prefix: '/' });
+}
